Add draft include header
This is meant to be a first draft of the header file. I've given some thought
to what the API will look like, but haven't set it in stone. Of course, this
is not even 0.x yet. Nothing works.

Have decided that the library prefix will be Wg, meaning W~or~g, of course.
Also decided to go with a CamelCase for types and a pascalCase for members.

The idea is to map OpenGL's VAOs to WgObjects. Each has a WgBuffer for data.
The WgBuffer is equivalent to a VBO. Element Arrays are a little more
abstracted. The WgObject will have a list of WgPrimitives, each primitive has
a fixed-length array of indices to the Elements in the WgBuffer.